Feat: 유저 서비스 구현
Resolves: #447

+++ user-service/src/main/java/com/example/userservice/user/adapter/in/web/dto/UserPointsResponse.java
@@ -11,7 +11,7 @@ public record UserPointsResponse(
     public static UserPointsResponse from(UserBalanceResult result) {
         return UserPointsResponse.builder()
                 .userId(result.userId())
-                .availablePoints(result.availablePoints())
+                .availablePoints(result.availablePoints().longValue())
                 .build();
     }
 }
